Replaces WeakValueDict IDProxy registry with a synchronized version
We've been synchronizing access in getproxy, but I think I overlooked
that when our IDProxys are garbage collected, a callback occurs which
removes the relevant entries from the WeakValueDict. That callback is
not synchronized, and so it appears to be possible for a race condition
to arise between getproxy and the garbage collector. This change
addresses the problem.