Remove unnecessary import.  (Fixes Python 3.3 reference problem.)
The import sys line is unnecessary because we have already imported
it at the global scope earlier in the file. This prevents a very
strange reference leak which I do not quite understand. The leak
manifests itself in Python 3.3 as a failing test (test_attrs_data.py:
TestTypes.test_complex).

The bug was tracked down by turning on `gc.set_debug(gc.DEBUG_LEAK)`
and noticing an entry in `gc.garbage`::

   ImportError("No module named 'h5py.sys'",)

Certainly this should not have caused an import error, so it is
potentially a bug in Cython. More investigation will be required.

Closes #275.