removed properties related to trace file because, to quote the documentation:
"The trace file feature is unsupported unless used at the direction of
The HDF Group. It is intended to allow The HDF Group to collect a
trace of cache activity in cases of occult failures and/or poor
performance seen in the field, so as to aid in reproduction in the
lab. If you use it absent the direction of The HDF Group, you are on
your own."