use dtype.metadata for special dtypes
Originally, there was no way to introduce private data into
numpy dtype. So, h5py was abusing the fields of non-compound
datatypes to magically introduce metadata. This was regularly
becoming a problem since numpy did not expect to see this.

Since several years now, numpy has added a metadata-field to
dtype. Let's use it. This is what it is for.