Conditional declaration of SWMR functions in the HL Dataset class
Changing the Dataset class so that the refresh and flush functions only
exist in the class if the SWMR feature is available in the HDF5 library.
This is done by testing for the presence of refresh and flush functions
in the h5d.DatasetID class - which in turn are conditional on the HDF5
library version.