Package Config: Fix Detection
Fix the detection of `pkgconfig` to avoid fallbacks to
HDF `1.8.4` or non-detection on default installs on Ubuntu/Debian
based systems.

This is a fixed version of PR #707 by @asford with the feedback
provided by @andreabedini.

The main problem is the mismatch of `list`
[append vs. extend](http://stackoverflow.com/questions/252703/append-vs-extend)
:
`pkgconfig.parse("hdf5")` returns a dictionary of `set`s of which we
take the set of `library_dirs` which should be extending the libdirs
list element-wise (and *not* "adding a `set` object to a list"
via `append`).