Avoid unaligned memory accesses in conversion functions
No guarantees are provided for the alignment of the buffers used for the
various conversion functions, and in the case of packed structs (easily
creatable with np.dtype, and in fact the test suite does so on the odd
occasion) they definitely can be unaligned. Therefore exclusively use
memcpy for multi-byte accesses to these buffers. The test suite now
passes on Linux sparc64, which was previously crashing due to unaligned
accesses.