Prefer custom HDF5 over pkg-config/fallback paths
When specifying a custom HDF5 location, alternative locations for the
library from system pkg-config files and the hardcoded fallback paths
can interfere with proper operation of the specified library.

If a location is specified, avoid invoking pkg-config and the fallback
paths to honor the user intent.

This addresses GitHub issue #946.