FIX: Explicitly cast float16 values in make_new_dataset
If the user:
 - passes in non-float16 data
 - specifies a dtype of float16

they may hit a suspected bug in the HDF5 conversion code that
incorrectly converts particular numbers (when the mantissa of the
float16 value is 0 and the float64 value is a fraction of float16 eps
closer to 0 than the float16 number).

This identifies that case and explicitly casts to 'f2' in python
before handing the values off to hdf5.

closes #819