Further simplify `_get_available_ftypes`
As there is no need to handle repeat types, cutdown the `for`-loop at
the end collecting type information about the floating precision types
available in `itemsize` sorted order to the bare minimum. Also type all
arguments explicitly so that Cython is able to make some optimizations
for these types. Makes the code a bit more understandable as well.