Simply use `cdef` with `_get_available_ftypes`
The `_get_available_ftypes` function is an internal helper function used
in `h5t.pyx` to initialize `_available_ftypes`. As such there is no need
for it to be anything other than a straight `cdef` function as it will
not be used outside of this Cython file. This should improve its
performance further and simplify the C code generated for it.