Drop explicit sorting of types by itemsize
NumPy provides a consistent ordering of `typecodes`. The floating
typecodes in particular have not changed order in the past 8 years
(since NumPy 1.6.0), which was only to fix the typecode for half
precision floats (not change the order). As these are already in order
of increasing itemsize and NumPy has left them in that order for a very
long time, just trust that NumPy will provide them in a reasonable order
and maintain that order. This simplifies the code in
`_get_available_ftypes` by avoiding an extra `for`-loop, which improve
performance and readability.