Fix for zero length collective IO (#965)
Collective IO where one rank attempts to read or write a zero length
slice causes the a hang in an underlying collective call. This fix
ensures that all ranks participate in any collective IO operation even
if some operations are on zero length slices.

This also includes some basic unit testing for collective IO that tests
for this problem.