Create an example of storing units in HDF5
This example script demonstrates how to store unit information attached to an array of numbers using HDF5 in such a way that the original data can be recovered upon reading the file. It uses the Unyt package (https://github.com/yt-project/unyt) for handling units, but presumably similar methods would work for people using astropy.units or Pint. It relies on Unyt being able to be able to parse its own string-ified versions of the units attached correctly, but as it should hopefully create such strings consistently it should work for the majority of use cases.