update cython defs to v1.6 compatibility in v1.12
HDF5 v1.12 provides a compatibility feature, however it seems to not
have covered everything, i.e. the compatibility-v1.6 is not the same as
v1.6, so fixup the differences.

Note: with HDF5 v1.12.0, a further patch is required to H5version.h
(generated from H5vers.txt), to fix missing functions/types in the v1.6
compatibility spec (not sure whether it's a bug in upstream or not; it
seems that the information about the API version in which a function is
introduced might be inaccurate).

Fixes issues in #1504 and PR #1506.