TST: Skip problematic datatypes based on their length
Filter datatypes for tests based on their length rather than keeping
a list of illegal platform+type combinations.  This is more robust
than the previous solution and fixes the test failures with new numpy
versions (1.19.5 is the oldest I have tested) that declare 'float96'
and 'complex192' on i686 rather than 'float128' and 'complex256'.

This fixes the following test failure:

    E       AttributeError: module 'numpy' has no attribute 'float128'

plus, later on (again):

    E   TypeError: Illegal length 24 for complex dtype