FIX: memory leak in fast-reader Reader.read
Previously we were creating but never destroying memory space objects.  This
changes the code to reuse the `SpaceID` class which will take care of making
the c level lifeccyle track the Python level life cycle.

This bug was exposed by fixing a different bug in #1944 that prevented this
code from ever being run.

Closes #1975