Check string pointer in string fillvalue getter
Based on the generated cython code, we call PyBytes_FromString on the
pointer we receive back from H5Pget_fill_value. As it appears the
pointer either can remain null, or null is returned for the empty
string, and PyBytes_FromString explicitly does not handle null, and is
called out as not checking it, it seems it's possible to segfault python
if we pass in a null pointer. This lacks a test, as I have not yet found
a reproducible one yet.