Try to fix ARM wheel upload error
It looks like the ARM wheels have been failing for ~7 months now (based
on the uploaded wheels). This does the following to address that:
 * Make the script set -e, so that it errors if any command errors.
 * Add the missing conda-package-handling to hopefully let the script
   pass.

It's possible that this setup no longer works, and we need to use an
alternative to anaconda-client to upload wheels (but it looks like other
projects have no issues uploading, so we'll need to see what they do).