Include Blosc2 dependencies in Tox test environment configuration
The names are `blosc2-deps` and `blosc2-mindeps`.  For instance, it should be
possible to run the normal tests plus Blosc2-specific tests with an
environment name like `test-py310-deps-blosc2-deps`.

Please note that normal dependencies include `blosc>=2.3.0` but minimum ones
include `blosc2==2.3.1`.  This is to avoid the latter to specifically install
2.3.1 (whose PyPI wheels were yanked), while the former should still work and
install a non-yanked wheel, while conveying the idea that the code of version
2.3.0 is perfectly ok (though its wheels may be broken).