Clarify sort order
Sort order is not lexicographic on link names including the '/' separator.

More precisely link names are first splitted at the path separator, and the
resulting tuples are sorted lexicographically. To sort the tuples, part names
are ordered lexicographically.
In python syntax, if `l` is the list of link names, the iteration is
sorted according to `l.sort(key=lambda x: x.split('/'))`