Lazily load dcpl: it's not needed for reading
Datasets have a `_dcpl` (data creation property list)
field. For virtual Datasets it can be very expensive to
load this list since it contains the entire virtual to
underlying source mapping. Since this field is only
actually needed for *writing* to the Dataset this commit
defers loading the `_dcpl`.

This improves performance for opening a virtual Dataset
like one created in `versioned_hdf5` quite substantially.
Previously:
```
In [3]: with tempfile.TemporaryDirectory() as d:
   ...:     with h5py.File(f'{d}/data.h5', 'w') as f:
   ...:         vf = VersionedHDF5File(f)
   ...:         with vf.stage_version('r0') as sv:
   ...:             sv.create_dataset('values', data=np.random.rand(365, 12345), chunks=(10, 100))
   ...:     with h5py.File(f'{d}/data.h5', 'r') as f:
   ...:         vf = VersionedHDF5File(f)
   ...:         cv = vf[vf.current_version]
   ...:         def open_dataset():
   ...:             values = cv['values']
   ...:         print(timeit.timeit(open_dataset, number=100))
   ...:
3.910119968932122
```
With the changes in this commit:
```
In [5]: with tempfile.TemporaryDirectory() as d:
   ...:     with h5py.File(f'{d}/data.h5', 'w') as f:
   ...:         vf = VersionedHDF5File(f)
   ...:         with vf.stage_version('r0') as sv:
   ...:             sv.create_dataset('values', data=np.random.rand(365, 12345), chunks=(10, 100))
   ...:     with h5py.File(f'{d}/data.h5', 'r') as f:
   ...:         vf = VersionedHDF5File(f)
   ...:         cv = vf[vf.current_version]
   ...:         def open_dataset():
   ...:             values = cv['values']
   ...:         print(timeit.timeit(open_dataset, number=100))
   ...:
1.0834132973104715
```

Note that the performance here is still pretty poor since
there are some performance issues in the underlying HDF5
library implementation as well which also copy the entire
virtual "layout" when a Dataset is accessed. I also hope to
fix these issues in HDF5 itself.