Make phil Lock handling fork-safe
When we use multiple threads to read
via `h5py` with forked subprocesses we
can sometimes observe the subprocess hanging.
This happens because in the parent process one
of the threads currently holds the `phil` Lock
and when we fork only the current thread is kept
in the child process.
If the child process now tries to acquire the
`phil` Lock it will hang forever.

To work around this issue we explicitly acquire
the `phil` Lock in the thread which is about to
fork *before* forking and then release it again
*after* forking in both the parent and the child
process.

There are three scenarios:
1. no thread currently holds the `phil` Lock: in
   this case it is just acquired for the duration of
   the fork and immediately released after
2. the current thread holds the `phil` Lock: since
   the lock is reentrant this basically just increments
   the lock count before forking and decrements it
   after the fork
3. a different thread holds the `phil` Lock: in this
   case we would wait until the lock is released by
   the other thread and only then fork.

Closes #2584